+++ model/mall/response/mall_shop_cart.go
@@ -11,5 +11,5 @@ type CartItemResponse struct {
 
 	GoodsCoverImg string `json:"goodsCoverImg"`
 
-	SellingPrice int `json:"sellingPrice"`
+	SellingPrice float32 `json:"sellingPrice"`
 }
